fix: fix Error creating bean with name 'eventServiceImpl': Lookup method resolution failed

+++ stats-server/stats-server-dto/src/main/java/ru/practicum/HitDto.java
@@ -13,5 +13,5 @@ public class HitDto {
 
     String uri;
 
-    Long hits;
+    Integer hits;
 }
